requirements.txt
psycopg2-binary
google-cloud-bigquery
prettytable
